feat: add group chat support with full CRUD operations and UI components
Implement comprehensive group chat functionality including:

Backend (Rust):
- Add group API endpoints for create, update, invite members, and leave group
- Consolidate sync operations into incremental_operations.rs module
- Add RippleWsSyncHandler for WebSocket-based synchronization
- Add command message type support in protobuf schema
- Extend store engine with group-related operations

Frontend (Vue):
- Add group management dialogs (Create, Edit, Invite, Leave, View Members)
- Add GroupChatHeader component for group chat interface
- Add CommandMessage component for system/command messages
- Implement useGroupActions and useGroupsDisplay composables
- Update ChatView and PeopleView to support group conversations

+++ src-tauri/src/errors.rs
@@ -15,7 +15,7 @@ pub enum CommandError {
     #[error(transparent)]
     AnyhowError(#[from] anyhow::Error),
     #[error("Ripple API error: API `{0}`, status code {1}, message: {2}")]
-    RippleAPIError(String, i64, String),
+    RippleAPIError(String, i32, String),
 }
 
 impl serde::Serialize for CommandError {
